fix: cleaned up a lot of diff test code.
