+++ StockAnalysis.Tests/Diff/analysis.html
@@ -1,0 +1,13 @@
+<html>
+<body>
+<h2>New positions</h2>
+<table border=1 frame=void rules=rows,columns>
+<tr><th>Company name</th><th>ticker</th><th>#shares</th><th>weight(%)</th></tr>
+<tr><td>TESLA INC</td><td>TSLA</td><td>4028071</td><td>9.83</td></tr>
+<tr><td>COINBASE GLOBAL INC -CLASS A</td><td>COIN</td><td>2630233</td><td>8.88</td></tr>
+<tr><td>ROKU INC</td><td>ROKU</td><td>8941303</td><td>7.96</td></tr>
+</table>
+<h2>Increased positions</h2>
+<h2>Reduced positions</h2>
+</body>
+</html>
