+++ StockAnalysis.Tests/DiffTests/DiffStoreTests/DiffStoreCreatorTests.cs
@@ -1,6 +1,6 @@
 using StockAnalysis.Diff.Store;
 
-namespace StockAnalysisTests.DiffTests;
+namespace StockAnalysisTests.DiffTests.DiffStoreTests;
 
 public class DiffStoreCreatorTests
 {
