+++ StockAnalysis.Tests/DiffTests/DiffStoreTests/HtmlDiffStoreTests.StoreDiff_WhenCalledRight_ShouldBeRightContentInCreatedFile.approved.html
